requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
DateTime==5.5
et-xmlfile==1.1.0
fonttools==4.54.1
idna==3.10
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.0.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
pyTelegramBotAPI==4.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
telebot==0.0.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yadisk==3.1.0
zipp==3.20.2
zope.interface==7.1.0
